requirements.txt
Pillow==9.5.0
torch==1.13.0
torchvision==0.14.0
tqdm==4.64.1
